New implementation of AMG using CSRM matricies
New implementation rewrites the way that the A matrix is stored.  This
results in a speed up for non AMG iterations.  For iterations with
AMG there is a decrease in speed compared to the previous method.
Commiting as it is working as intended, will need to investigate
which component of the AMG is slow.  My guess is it is the generation of
RAP which requires two Sparse*Dense operations.  May also be the group
building.  Will test optimizations of both.