Efficiency improvements for new AMG method
Rewrote large portions of the matrix multiplication for the sparse
sparse matrix products to use row-wise product.  Results in a
significant speed up in computation time for AMG.  Also converted the
Restriction and Prolongation matrices into CSRM as well to help facilitate
this.  Increased overhead means little to no improvement on very coarse grids
(only a couple thousand cells).  In that case no AMG may still be faster.
For grids of a few thousand + significant speedups can be seen both in
a large decrease in the number of coarse (AMG level 0) smoothing loops,
and a decrease in the number of total iterations to reach a converged
solution.  This results in a significant speed up in computation time.

STILL TODO: (1)Clean up miscellaneous commented code that remains as a
result of switching to new methods.  This was kept for debugging purposes.
(2)Add documentation to the numerous subroutines that have been created
to facilitate the new AMG for easier reading in the future. (3)Check
for memory leaks.  I used pointers in a few spots and I'm not 100% sure
I've cleaned everything up properly.  I might see if I can rewrite
this without pointers at some point to limit the risk of weird memory
errors. (4)Come up with a next feature to implement! I'd say AMG is
pretty well working, time to start on the next thing... (Full NS?!?!)