Added (hopefully) working Weiss-Smith low-Mach preconditioning.
Some weridness in the simple test case but a considerable speed up
in convergence for more complex cases.  It appears to be working so
I'm pushing.  Still need to do a decent amount of clean up to take
care of all the messy debugging code that is still there.  I'll do
that at some point...